+++ CompetitiveStatTracker/SmiteDataSet.vb
@@ -1,3 +1,9 @@
 ﻿Partial Class SmiteDataSet
 End Class
 
+
+Namespace SmiteDataSetTableAdapters
+
+    Partial Public Class GodWinsTableAdapter
+    End Class
+End Namespace